+++ vision/packages/object_detector_2d/scripts/detectors/utils.py
